extract systemInfo and random functions to their own files
