+++ random.js
@@ -1,0 +1,7 @@
+function random() {
+  return Math.floor(Math.random() * 10) + 1;
+}
+
+module.exports = {
+  random,
+};
